Replace range slier with edit text

+++ app/src/main/java/com/openclassrooms/realestatemanager/mappers/propertiesToFilterMapper.kt
@@ -6,26 +6,26 @@ import com.openclassrooms.realestatemanager.models.UserData
 
 
 fun propertiesToFilterMapper(properties: List<Property>, userData: UserData): PropertyFilter { // TODO User data
-    var minPrice = 0.0
-    var maxPrice = 0.0
-    var minSurface = 0.0
-    var maxSurface = 0.0
+    var minPrice = 0L
+    var maxPrice = 0L
+    var minSurface = 0L
+    var maxSurface = 0L
 
     for(property in properties){
         if(property.price != null) {
             if (property.price!! < minPrice) {
-                minPrice = property.price!!
+                minPrice = property.price!!.toLong()
             }
             if (property.price!! > maxPrice) {
-                maxPrice = property.price!!
+                maxPrice = property.price!!.toLong()
             }
         }
         if(property.surface != null) {
             if (property.surface!! < minSurface) {
-                minSurface = property.surface!!
+                minSurface = property.surface!!.toLong()
             }
             if (property.surface!! > maxSurface) {
-                maxSurface = property.surface!!
+                maxSurface = property.surface!!.toLong()
             }
         }
     }
